Chart display done.

+++ lib/widgets/new_transaction.dart
@@ -16,7 +16,7 @@ class _NewTransactionState extends State<NewTransaction> {
 
   void onSubmit() {
     final enteredTitle = titleController.text;
-    final enteredAmount = int.parse(amountController.text);
+    final enteredAmount = double.parse(amountController.text);
 
     if (enteredTitle.isEmpty || enteredAmount <= 0) {
       return;
